+++ email-verifier/src/index.ts
@@ -2,11 +2,12 @@ import * as bodyParser from 'body-parser';
 import * as dotenv from 'dotenv';
 import * as express from 'express';
 import { filterXSS } from 'xss';
-import { single } from './api';
-import { validateSinglePhone } from './apiPhoneVerifier';
+import { bulk, single } from './api';
+import { validateBulkPhones, validateSinglePhone } from './apiPhoneVerifier';
 import { connect } from './connection';
-import { initConsumer } from './messageBroker';
-import { debugBase, debugRequest } from './utils';
+import './cronJobs/verifier';
+import { initRedis } from './redisClient';
+import { debugBase, debugCrons, debugRequest } from './utils';
 
 // load environment variables
 dotenv.config();
@@ -22,11 +23,24 @@ app.post('/verify-single', async (req, res, next) => {
   const { email } = req.body;
 
   try {
-    const status = await single(email, true);
+    const result = await single(email);
 
-    return res.json({ status });
+    return res.json(result);
   } catch (e) {
     return next(new Error(e));
+  }
+});
+
+app.post('/verify-bulkEmails', async (req, res, next) => {
+  debugRequest(debugBase, req);
+
+  const { emails, hostname } = req.body;
+
+  try {
+    const result = await bulk(emails, hostname);
+    return res.json({ emails: result });
+  } catch (e) {
+    return next(e);
   }
 });
 
@@ -36,11 +50,25 @@ app.post('/verify-singlePhone', async (req, res, next) => {
   const { phone } = req.body;
 
   try {
-    const result = await validateSinglePhone(phone, true);
+    const result = await validateSinglePhone(phone);
 
-    return res.json({ result });
+    return res.json(result);
   } catch (e) {
     return next(new Error(e));
+  }
+});
+
+app.post('/verify-bulkPhones', async (req, res, next) => {
+  debugRequest(debugBase, req);
+
+  const { phones, hostname } = req.body;
+
+  try {
+    const result = await validateBulkPhones(phones, hostname);
+
+    return res.json({ phones: result });
+  } catch (e) {
+    return next(e);
   }
 });
 
@@ -55,8 +83,13 @@ app.use((error, _req, res, _next) => {
 const { PORT } = process.env;
 
 app.listen(PORT, async () => {
-  await initConsumer();
   await connect();
-
+  initRedis();
   debugBase(`Email verifier server is running on port ${PORT}`);
+});
+
+const { PORT_CRONS = 4700 } = process.env;
+
+app.listen(PORT_CRONS, () => {
+  debugCrons(`Cron Server is now running on ${PORT_CRONS}`);
 });