+++ src/data/schema/integration.ts
@@ -173,29 +173,5 @@ export const mutations = `
     attachments: [JSON]
   ): JSON
 
-  integrationAddExchangeAccount(
-    email: String!
-    password: String!
-    host: String!
-    username: String
-    kind: String
-  ): JSON
-
-  integrationAddImapAccount(
-    email: String!
-    password: String!
-    imapHost: String!
-    imapPort: Int!
-    smtpHost: String!
-    smtpPort: Int!
-    kind: String!
-  ): JSON
-
-  integrationAddMailAccount(
-    email: String!
-    password: String!
-    kind: String!
-  ): JSON
-
   integrationsUpdateConfigs(configsMap: JSON!): JSON
 `;