+++ src/workers/messageBroker.ts
@@ -2,7 +2,8 @@ import * as amqplib from 'amqplib';
 import * as dotenv from 'dotenv';
 import { RABBITMQ_QUEUES } from '../data/constants';
 import { debugWorkers } from '../debuggers';
-import { receiveImportCancel, receiveImportRemove, receiveImportXls } from './utils';
+import { addToArray, getArray, removeFromArray } from '../redisClient';
+import { generateUid, receiveImportCancel, receiveImportRemove, receiveImportXls } from './utils';
 
 dotenv.config();
 
@@ -36,11 +37,26 @@ export const initConsumer = async () => {
 
     channel.consume(RABBITMQ_QUEUES.RPC_API_TO_WORKERS, async msg => {
       if (msg !== null) {
+        const response = { status: 'success', data: {}, errorMessage: '' };
+
+        const activeWorkers = await getArray('active_workers');
+
+        if (activeWorkers.length > 3) {
+          response.status = 'error';
+          response.errorMessage = 'Please white for a while';
+
+          return channel.sendToQueue(msg.properties.replyTo, Buffer.from(JSON.stringify(response)), {
+            correlationId: msg.properties.correlationId,
+          });
+        }
+
+        const uid = generateUid();
+
+        await addToArray('active_workers', uid);
+
         debugWorkers(`Received rpc queue message ${msg.content.toString()}`);
 
         const content = JSON.parse(msg.content.toString());
-
-        const response = { status: 'success', data: {}, errorMessage: '' };
 
         try {
           response.data =
@@ -49,6 +65,8 @@ export const initConsumer = async () => {
           response.status = 'error';
           response.errorMessage = e.message;
         }
+
+        await removeFromArray('active_workers', uid);
 
         channel.sendToQueue(msg.properties.replyTo, Buffer.from(JSON.stringify(response)), {
           correlationId: msg.properties.correlationId,