+++ src/workers/utils.ts
@@ -2,12 +2,15 @@ import * as dotenv from 'dotenv';
 import * as mongoose from 'mongoose';
 import * as os from 'os';
 import * as path from 'path';
+import * as XlsxStreamReader from 'xlsx-stream-reader';
 import { checkFieldNames } from '../data/modules/fields/utils';
+import { deleteFileAWS, s3Stream } from '../data/utils';
 import { ImportHistory } from '../db/models';
 import ImportHistories from '../db/models/ImportHistory';
-import { debugImport } from '../debuggers';
+import { debugImport, debugWorkers } from '../debuggers';
 
 const { MONGO_URL = '' } = process.env;
+
 export const connect = () => mongoose.connect(MONGO_URL, { useNewUrlParser: true, useCreateIndex: true });
 
 dotenv.config();
@@ -130,43 +133,125 @@ export const receiveImportCancel = () => {
   return { status: 'ok' };
 };
 
-export const receiveImportXls = async (content: any) => {
-  const { type, fieldNames, scopeBrandIds, user, datas } = content;
+const readXlsFile = async (fileName: string): Promise<{ fieldNames: string[]; datas: any[] }> => {
+  return new Promise(async (resolve, reject) => {
+    let rowCount = 0;
 
-  if (datas.length === 0) {
-    throw new Error('Please import at least one row of data');
-  }
+    const usedSheets: any[] = [];
 
-  const properties = await checkFieldNames(type, fieldNames);
+    const xlsxReader = XlsxStreamReader();
 
-  const importHistory = await ImportHistory.create({
-    contentType: type,
-    total: datas.length,
-    userId: user._id,
-    date: Date.now(),
+    const errorCallback = error => {
+      reject(new Error(error.code));
+    };
+
+    try {
+      const stream = await s3Stream(fileName, errorCallback);
+
+      stream.pipe(xlsxReader);
+
+      xlsxReader.on('worksheet', workSheetReader => {
+        if (workSheetReader > 1) {
+          return workSheetReader.skip();
+        }
+
+        workSheetReader.on('row', row => {
+          if (rowCount > 100000) {
+            return reject(new Error('You can only import 100000 rows one at a time'));
+          }
+
+          if (row.values.length > 0) {
+            usedSheets.push(row.values);
+            rowCount++;
+          }
+        });
+
+        workSheetReader.process();
+      });
+
+      xlsxReader.on('end', () => {
+        const compactedRows: any = [];
+
+        for (const row of usedSheets) {
+          if (row.length > 0) {
+            row.shift();
+
+            compactedRows.push(row);
+          }
+        }
+
+        const fieldNames = usedSheets[0];
+
+        // Removing column
+        compactedRows.shift();
+
+        return resolve({ fieldNames, datas: compactedRows });
+      });
+
+      xlsxReader.on('error', error => {
+        return reject(error);
+      });
+    } catch (e) {
+      reject(e);
+    }
   });
+};
 
-  const results: string[] = splitToCore(datas);
+export const receiveImportXls = async (content: any) => {
+  try {
+    const { fileName, type, scopeBrandIds, user } = content;
 
-  const workerFile =
-    process.env.NODE_ENV === 'production'
-      ? `./dist/workers/bulkInsert.worker.js`
-      : './src/workers/bulkInsert.worker.import.js';
+    const { fieldNames, datas } = await readXlsFile(fileName);
 
-  const workerPath = path.resolve(workerFile);
+    if (datas.length === 0) {
+      throw new Error('Please import at least one row of data');
+    }
 
-  const percentagePerData = Number(((1 / datas.length) * 100).toFixed(3));
+    const properties = await checkFieldNames(type, fieldNames);
 
-  const workerData = {
-    scopeBrandIds,
-    user,
-    contentType: type,
-    properties,
-    importHistoryId: importHistory._id,
-    percentagePerData,
-  };
+    const importHistory = await ImportHistory.create({
+      contentType: type,
+      total: datas.length,
+      userId: user._id,
+      date: Date.now(),
+    });
 
-  await createWorkers(workerPath, workerData, results);
+    const results: string[] = splitToCore(datas);
 
-  return { id: importHistory.id };
+    const workerFile =
+      process.env.NODE_ENV === 'production'
+        ? `./dist/workers/bulkInsert.worker.js`
+        : './src/workers/bulkInsert.worker.import.js';
+
+    const workerPath = path.resolve(workerFile);
+
+    const percentagePerData = Number(((1 / datas.length) * 100).toFixed(3));
+
+    const workerData = {
+      scopeBrandIds,
+      user,
+      contentType: type,
+      properties,
+      importHistoryId: importHistory._id,
+      percentagePerData,
+    };
+
+    await createWorkers(workerPath, workerData, results);
+
+    await deleteFileAWS(fileName);
+
+    return { id: importHistory.id };
+  } catch (e) {
+    debugWorkers(e.message);
+    throw e;
+  }
+};
+
+export const generateUid = () => {
+  return (
+    '_' +
+    Math.random()
+      .toString(36)
+      .substr(2, 9)
+  );
 };